SQL for bulk loading data
Turns out the flight dataset is actually huge... somewherea round 5 million rows. On my computer the bulk load took about 10 minutes to load into the database. I'm also off campus using a VPN though so there is probably some more network latency than if you are on campus and try to do it. Anyways, it can probably use some cleaning in the database, but it works for now.